refactor: split event and action tables into separate files
